* hamt.c: Use uintptr_t to correctly cast to integer from pointer.  On some platforms (notably Win64), unsigned long is not big enough to hold a pointer. * Makefile.am, configure.ac: Use ax_create_stdint_h to get us uintptr_t. * ax_create_stdint_h.m4: Implementation of ax_create_stdint_h autoconf macro from http://ac-archive.sourceforge.net/guidod/ax_create_stdint_h.html
svn path=/trunk/yasm/; revision=1223

+++ libyasm/hamt.c
@@ -35,6 +35,7 @@
 
 #include "coretype.h"
 #include "hamt.h"
+#include "_stdint.h"			/* for uintptr_t */
 
 typedef struct HAMTEntry {
     STAILQ_ENTRY(HAMTEntry) next;	/* next hash table entry */
@@ -58,14 +59,14 @@ struct HAMT {
  * 4 or 2-byte aligned (as it uses the LSB of the pointer variable to store
  * the subtrie flag!
  */
-#define IsSubTrie(n)		((unsigned long)((n)->BaseValue) & 1)
+#define IsSubTrie(n)		((uintptr_t)((n)->BaseValue) & 1)
 #define SetSubTrie(h, n, v)	do {				\
-	if ((unsigned long)(v) & 1)				\
+	if ((uintptr_t)(v) & 1)				\
 	    h->error_func(__FILE__, __LINE__,			\
 			  N_("Subtrie is seen as subtrie before flag is set (misaligned?)"));	\
-	(n)->BaseValue = (void *)((unsigned long)(v) | 1);	\
+	(n)->BaseValue = (void *)((uintptr_t)(v) | 1);	\
     } while (0)
-#define GetSubTrie(n)		(HAMTNode *)((unsigned long)((n)->BaseValue)&~1UL)
+#define GetSubTrie(n)		(HAMTNode *)((uintptr_t)((n)->BaseValue)&~1UL)
 
 static unsigned long
 HashKey(const char *key)
